app/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
python-dotenv==0.18.0
psycopg2-binary==2.8.6
drf-yasg==1.20.0
django-filter==2.4.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
dj-database-url==0.5.0
